Subprogramの実装をWindowsとUnixでファイルを分けた。
